feat: add edit command for image properties
